OrderTicket optional titlebar
